# docs/swagger.yaml
definitions:
...
  model.CategoryResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  model.Pagination:
    properties:
      page:
        type: integer
      size:
        type: integer
      totalItem:
        type: integer
      totalPage:
        type: integer
    type: object
  model.PostResponse:
    properties:
      categoryID:
        type: integer
      content:
        type: string
      createdAt:
        type: integer
      id:
        type: integer
      summary:
        type: string
      thumbnail:
        type: string
      title:
        type: string
      updatedAt:
        type: integer
      userID:
        type: integer
    type: object
  model.PostResponseWithPreload:
    properties:
      category:
        $ref: '#/definitions/model.CategoryResponse'
      content:
        type: string
      createdAt:
        type: integer
      id:
        type: integer
      summary:
        type: string
      thumbnail:
        type: string
      title:
        type: string
      updatedAt:
        type: integer
      user:
        $ref: '#/definitions/model.UserResponse'
      viewCount:
        type: integer
    type: object
...
  model.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phoneNumber:
        type: string
      profilePicture:
        type: string
      role:
        type: string
    type: object
...
  utility.PaginationResponse:
    properties:
      data: {}
      pagination: {}
    type: object
  utility.ResponseError:
    properties:
      error:
        type: string
    type: object
  utility.ResponseSuccess:
    properties:
      data: {}
    type: object
info:
  contact: {}
paths:
...
  /api/post:
    get:
      description: Search posts with various filtering options
      parameters:
      - default: 0
        description: Page number
        in: query
        name: page
        type: integer
      - default: 5
        description: Page size
        in: query
        name: size
        type: integer
      - default: 0
        description: User ID
        in: query
        name: userID
        type: integer
      - description: Title search query
        in: query
        name: title
        type: string
      - description: User name search query
        in: query
        name: userName
        type: string
      - description: Summary search query
        in: query
        name: summary
        type: string
      - description: Category name search query
        in: query
        name: categoryName
        type: string
      - description: 'Sort by: view_count, -view_count, created_at, -created_at'
        in: query
        name: sort
        type: string
      - description: Filter posts published after this date (timestamp)
        in: query
        name: startDate
        type: integer
      - description: Filter posts published before this date (timestamp)
        in: query
        name: endDate
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utility.PaginationResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.PostResponseWithPreload'
                  type: array
                pagination:
                  items:
                    $ref: '#/definitions/model.Pagination'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Search posts
      tags:
      - Post
    post:
      consumes:
      - multipart/form-data
      description: Create a new post with the given details
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Post Title
        in: formData
        name: title
        required: true
        type: string
      - description: Post Summary
        in: formData
        name: summary
        required: true
        type: string
      - description: Post Content
        in: formData
        name: content
        required: true
        type: string
      - description: User ID
        in: formData
        name: userID
        type: integer
      - description: Category ID
        in: formData
        name: categoryID
        required: true
        type: integer
      - description: Post Thumbnail
        in: formData
        name: thumbnail
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utility.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.PostResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Create a new post
      tags:
      - Post
  /api/post/{id}:
    delete:
      description: Delete a post by its ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utility.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Delete a post
      tags:
      - Post
    get:
      description: Retrieve a specific post by its ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utility.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.PostResponseWithPreload'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Get a post by ID
      tags:
      - Post
    put:
      consumes:
      - multipart/form-data
      description: Update an existing post's details
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Post Title
        in: formData
        name: title
        required: true
        type: string
      - description: Post Summary
        in: formData
        name: summary
        required: true
        type: string
      - description: Post Content
        in: formData
        name: content
        required: true
        type: string
      - description: User ID
        in: formData
        name: userID
        type: integer
      - description: Category ID
        in: formData
        name: categoryID
        required: true
        type: integer
      - description: Post Thumbnail
        in: formData
        name: thumbnail
        type: file
      - description: Delete Thumbnail
        in: formData
        name: deleteThumbnail
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utility.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.PostResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Update an existing post
      tags:
      - Post
  /api/post/{id}/view:
    patch:
      description: Increment the view count for a specific post by its ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utility.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Increment post view count
      tags:
      - Post
...
swagger: "2.0"